// src/types/tributario.ts
export type Anexo = 'I' | 'II' | 'III' | 'IV' | 'V'
// ...
export interface ResultadoAliquota {
  rbt12: number
  faixa: number
  aliquotaNominal: number
  parcelaDeduzir: number
  aliquotaEfetiva: number
  dasAnual: number
  dasMensal: number
}